Make the ckanext-oaipmh extension return B2FIND records in Datacite
If applied, this commit will update the ckanext-oaipmh extension so that
it returns B2FIND records in Datacite 3.1 format. The new fields are defined
in oaipmh_server.py.

This work was done for the reason that B2Find data could be harvested by
OpenAire in Datacite format.

The side-effects are, that since Datacite requires DOI, files without DOI
can't be fetched.

The code is tested with dataset from Materials Cloud.